fix: Enable app working offline after quitting the app and re-open
Resolves: DCMAW-15928
